falta que las soluciones convergan

+++ pso/psoprueba/Particle.java
@@ -1,10 +1,10 @@
 public class Particle {
-    double[] position; //The position vector of this particle
+    int[] position; //The position vector of this particle
 	double fitness; //The fitness of this particle
 	double[] velocity; //The velocity vector of this particle
-	double[] personalBest; //Personal best of the particle
+	int[] personalBest; //Personal best of the particle
 
-	public Particle(double[] position, double[] velocity) {
+	public Particle(int[] position, double[] velocity) {
 		this.position = position;
 		this.velocity = velocity; 
 	}
